Fix issue running setup.py egg_info
Since the main __init__.py imports most of the modules in the project,
it was not a good idea to add the `__version__` rariable there. While
doing:

```
from ravenpackapi import __version__
```

imported everything else, so when we run

```
python setup.py egg_info
```

For example, we would get errors like `No module named 'retry'`.
By storing the variable in a separate module, we can do some "tricks" to
import it in the `setup.py` but skipping the `__init__.py`.